requirements.txt
aiomysql==0.2.0
aioredis==1.3.1
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
fastapi-cache==0.1.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
httptools==0.6.4
idna==3.10
jose==1.0.0
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
